requirements.txt
python-dotenv
SQLAlchemy
psycopg2-binary
dash
pandas
gunicorn
bcrypt